+++ components/molecules/Content.tsx
@@ -1,8 +1,14 @@
-import { ActionIcon, Text, createStyles, rem } from "@mantine/core";
+import styled from "@emotion/styled";
+import { ActionIcon, Text, Tooltip, createStyles, rem } from "@mantine/core";
 import { IoIosMore } from "react-icons/io";
 import { ToolbarMenu } from "../atoms/ToolbarMenu";
-import styled from "@emotion/styled";
 
+interface ContentProps {
+  no?: number;
+  verse?: string;
+  ayat?: string;
+  translation?: string;
+}
 const useStyles = createStyles((theme) => ({
   wrapper: {
     width: "100%",
@@ -10,6 +16,9 @@ const useStyles = createStyles((theme) => ({
     borderBottom: `${rem(1)} solid ${
       theme.colorScheme === "dark" ? theme.colors.dark[4] : theme.colors.gray[2]
     }`,
+    "&:last-child": {
+      borderBottom: "none",
+    },
   },
   container: {
     display: "flex",
@@ -36,7 +45,7 @@ const useStyles = createStyles((theme) => ({
 }));
 
 // gap: ${({ theme }) => theme.spacing.md};
-const ArabicContent = styled.div`
+const QuranContent = styled.div`
   display: flex;
   flex-direction: column;
   width: 100%;
@@ -50,46 +59,49 @@ const ArabicContent = styled.div`
   .quran-text {
     font-family: "Quran";
     font-size: ${rem(28)};
+    width: 100%;
+  }
+
+  .translation-text {
+    font-size: ${rem(16)};
+    width: 100%;
   }
 `;
 
-export const Content = () => {
-  const { classes, cx } = useStyles();
+export const Content = ({
+  no = 0,
+  verse = "1:1",
+  ayat = "ayat",
+  translation = "Terjemah",
+}: ContentProps) => {
+  const { classes } = useStyles();
   return (
     <div className={classes.wrapper}>
       <div className={classes.container}>
         <div className={classes.toolbar}>
-          <Text size="md">No: 1</Text>
-          <Text size="md">2:282</Text>
-          <ToolbarMenu>
-            <ActionIcon
-              variant="filled"
-              radius="xl"
-              size="md"
-              aria-label="Toolbar Menu"
-            >
-              <IoIosMore size="1.1rem" stroke="1.5" />
-            </ActionIcon>
+          <Text size="md">{`No: ${no}`}</Text>
+          <Text size="md">{verse}</Text>
+          <ToolbarMenu index={no}>
+            <Tooltip label="More" position="bottom" withArrow arrowSize={6}>
+              <ActionIcon
+                variant="filled"
+                radius="xl"
+                size="md"
+                aria-label="Toolbar Menu"
+              >
+                <IoIosMore size="1.1rem" stroke="1.5" />
+              </ActionIcon>
+            </Tooltip>
           </ToolbarMenu>
         </div>
-        <ArabicContent>
+        <QuranContent>
           <Text dir="rtl" className="quran-text" align="justify">
-            ۞ وَإِن كُنتُمْ عَلَىٰ سَفَرٍ وَلَمْ تَجِدُوا۟ كَاتِبًا فَرِهَـٰنٌ
-            مَّقْبُوضَةٌ ۖ فَإِنْ أَمِنَ بَعْضُكُم بَعْضًا فَلْيُؤَدِّ ٱلَّذِى
-            ٱؤْتُمِنَ أَمَـٰنَتَهُۥ وَلْيَتَّقِ ٱللَّهَ رَبَّهُۥ ۗ وَلَا
-            تَكْتُمُوا۟ ٱلشَّهَـٰدَةَ ۚ وَمَن يَكْتُمْهَا فَإِنَّهُۥٓ ءَاثِمٌ
-            قَلْبُهُۥ ۗ وَٱللَّهُ بِمَا تَعْمَلُونَ عَلِيمٌ
+            {ayat}
           </Text>
-          <Text align="justify">
-            Dan jika kamu dalam perjalanan sedang kamu tidak mendapatkan seorang
-            penulis, maka hendaklah ada barang jaminan yang dipegang. Tetapi,
-            jika sebagian kamu mempercayai sebagian yang lain, hendaklah yang
-            dipercayai itu menunaikan amanatnya (utangnya) dan hendaklah dia
-            bertakwa kepada Allah, Tuhannya. Dan janganlah kamu menyembunyikan
-            kesaksian, karena barangsiapa menyembunyikannya, sungguh, hatinya
-            kotor (berdosa). Allah Maha Mengetahui apa yang kamu kerjakan.
+          <Text className="translation-text" align="justify">
+            {translation}
           </Text>
-        </ArabicContent>
+        </QuranContent>
       </div>
     </div>
   );