+++ pages/akarkata.tsx
@@ -1,0 +1,7 @@
+import { NextPage } from "next";
+
+const AkarKata: NextPage = () => {
+  return <div>AkarKata</div>;
+};
+
+export default AkarKata;