+++ pages/cari/latin.tsx
@@ -1,0 +1,7 @@
+import { NextPage } from "next";
+
+const Latin: NextPage = () => {
+  return <div>Hello</div>;
+};
+
+export default Latin;