+++ pages/quantum.tsx
@@ -1,8 +1,0 @@
-import { NextPage } from "next";
-import React from "react";
-
-const Quantum: NextPage = () => {
-  return <div>Quantum</div>;
-};
-
-export default Quantum;